Add device and rename project
