+++ README.md
@@ -1,7 +1,7 @@
-# ULBLE Python Library
-Python library for interacting with Ultraloq locks via BLE.
+# UTECIO Python Library
+Python library for interacting with Utec devices and Ultraloq locks.
 
-This library is in initial development and will change rapidly. Development will be focused primarily on the UL1-BT and other MD5 Encryption Based locks initially with further locks added slowly as more functions are implimented.
+This library is in initial development and will change rapidly. Development will initially be focused on the UL1-BT and other MD5 Encryption Based BLE locks with further locks added slowly as more functions are implimented.
 
 If you have a specific lock request you can contribute at [buy me a coffee](https://www.buymeacoffee.com/maeneak) or [coindrop](https://coindrop.to/maeneak) to assist with purchasing of new locks for testing and accelerate development.
